login, rating and search functionality added

+++ controller/product.controller.js
@@ -1,9 +1,10 @@
-const { request, response } = require('express');
 const { MongoClient, ObjectId } = require('mongodb');
 const dbConnect = require('../config/dbconfig')
+
 const PRODUCT_COLLECTION = 'product';
 const CART_COLLECTION = 'cart';
 const ORDER_COLLECTION = 'order';
+const RATING_COLLECTION = 'rating';
 
 exports.getAllProduct = async (request, response) => {
     try {
@@ -20,7 +21,7 @@ exports.getAllProduct = async (request, response) => {
 
 exports.addToCart = async (request, response) => {
     const body = request.body;
-    body.userId = 1;
+    body.userId = request.userId;
 
     const { productId, productCatagory, userId } = body;
 
@@ -41,9 +42,10 @@ exports.addToCart = async (request, response) => {
 }
 
 exports.totalCartProduct = async (request, response) => {
+    const userId = request.userId;
     try {
         const collection = await dbConnect(CART_COLLECTION);
-        const countCartProduct = await collection.countDocuments();
+        const countCartProduct = await collection.countDocuments({ userId });
 
         return response.json({ countCartProduct })
     } catch (err) {
@@ -53,7 +55,7 @@ exports.totalCartProduct = async (request, response) => {
 }
 
 exports.allCartList = async (request, response) => {
-    const userId = 1;
+    const userId = request.userId;
 
     try {
         const collection = await dbConnect(CART_COLLECTION);
@@ -68,10 +70,11 @@ exports.allCartList = async (request, response) => {
 
 exports.deleteCart = async (request, response) => {
     const { cartId } = request.params;
+    const userId = request.userId;
 
     try {
         const collection = await dbConnect(CART_COLLECTION);
-        collection.deleteOne({ _id: new ObjectId(cartId) }).then(resp => {
+        collection.deleteOne({ _id: new ObjectId(cartId), userId }).then(resp => {
             return response.status(202).send();
         }).catch(err => response.status(500).send());
     } catch (err) {
@@ -81,7 +84,7 @@ exports.deleteCart = async (request, response) => {
 }
 
 exports.removeAllCart = async (request, response) => {
-    const userId = 1;
+    const userId = request.userId;
 
     try {
         const collection = await dbConnect(CART_COLLECTION);
@@ -95,8 +98,8 @@ exports.removeAllCart = async (request, response) => {
 }
 
 exports.checkout = async (request, response) => {
-    const userId = 1;
-    
+    const userId = request.userId;
+
     try {
         const cartCollection = await dbConnect(CART_COLLECTION);
         const cartDocument = await cartCollection.find({ userId }).toArray()
@@ -110,6 +113,39 @@ exports.checkout = async (request, response) => {
 
     } catch (err) {
         console.error(err);
+        response.status(500).send();
+    }
+}
+
+exports.addRating = async (request, response) => {
+    const userId = request.userId;
+    const userName = request.userName;
+    const product = request.body;
+    product.userId = userId;
+    product.userName = userName;
+    const { productId, category } = product;
+
+    try {
+        const collection = await dbConnect(RATING_COLLECTION);
+        await collection.insertOne(product);
+
+        return response.status(201).json({ productId, category });
+    } catch (error) {
+        console.error(error);
+        return response.status(500).send();
+    }
+}
+
+exports.retriveRating = async (request, response) => {
+    const { productId, category } = request.params;
+
+    try {
+        const collection = await dbConnect(RATING_COLLECTION);
+        const ratings = await collection.find({ productId, category }).toArray();
+
+        return response.json(ratings);
+    } catch (error) {
+        console.error(error);
         response.status(500).send();
     }
 }
